remoteproc/pruss_intc: deny module removal during active interrupts
The PRUSS INTC driver provides an irqchip interface for user clients
to use one of the 64 PRU system events as interrupt sources. There
are no references held though when a user client requests and is
using an interrupt. The PRUSS INTC module can still be removed at
present when there are active interrupts, and this can cause a kernel
crash/hang.

Fix this by holding a module reference count during the interrupt
setup phase. This will deny any attempts to remove the module. The
reference count is released when the interrupt is freed. This is
achieved through the irqchip's irq_request_resources() and
irq_release_resources() ops.

Fixes: 5b107622d1eb ("remoteproc/pruss: add platform drivers for PRUSS in TI SoCs")
Signed-off-by: Suman Anna <s-anna@ti.com>