TEMP: remoteproc/pru: enable ethernet fw for PRUSS1 on AM571x IDK
The PRU Ethernet functionality has only been enabled for the PRUSS2
instance on the AM57xx IDKs. The AM571x IDK v1.3a board is capable
of achieving PRU Ethernet functionality on PRUSS1 as well, but the
AM572x IDK still is not. The firmware data and capabilities are
maintained per SoC, so fixup the ethernet capability at runtime
for AM571x IDKs to enable the ethernet firmware to be loaded onto
the PRUSS1 PRU cores as well.

Signed-off-by: Suman Anna <s-anna@ti.com>