ti_config_fragments/defconfig_builder: Verify user choice
Verify that the user enters only a number.  If not
then indicate that on the screen and ask the user
to try again.

The user may try to type out the defconfig to build
because there is no specific instruction.  Updating the
builder request for the user to enter the number of the
defconfig to build.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Reported-by: Nishanth Menon <nm@ti.com>