ti_config_fragments: baseport: Enable AEABI
Though AEABI's kernel configuration option states "Use the ARM EABI to
compile the kernel" There is a lot of behind the scenes stuff going on
with that option.

The original objective is stated in the help text: "This option allows
for the kernel to be compiled using the latest ARM ABI (aka EABI). This
is only useful if you are using a user space environment that is also
compiled with EABI." However a bunch of optimizations do get enabled
with this option is enabled. Neon now becomes available, 64 bit
operations are optimized[1] and handles LPAE enabled systems more sanely,
etc..

NOTE: this also forces at least gcc 4.0.0 constraint on our build, but
since our current requirements are on 5.3.1, and AEABI seems standard
around the world, we should be benefiting from this change.

NOTE: The alternate approach of enabling AEABI based on SoC Kconfig
was discouraged[2] for Keystone (as example) and upstream is tending
towards a MULTIPLATFORM_TYPICAL configuration.

[1] http://marc.info/?t=135531876200008&r=1&w=2
[2] https://patchwork.kernel.org/patch/9148191/

Signed-off-by: Nishanth Menon <nm@ti.com>