ti_config_fragments: baseport: Enable Neon
All TI SoCs that are supported are v7 Cortex-A8/A9/A15 family of
processors that have Neon co-processors, so, enable NEON by default to
benefit hard float acceleration. See [1] for an sample description of
usage.

NOTE: The alternate approach of enabling NEON based on SoC Kconfig
was discouraged[2] for Keystone (as example) and upstream is tending
towards a MULTIPLATFORM_TYPICAL configuration.

[1] http://processors.wiki.ti.com/index.php/Cortex-A8
[2] https://patchwork.kernel.org/patch/9148191/

Signed-off-by: Nishanth Menon <nm@ti.com>