ti_config_fragments: Debug/Keystone: Enable HIGHMEM and debug options
Enable Highmem for Keystone and DRA7 platforms.
See[1] for more details, also notice the potential performance
impact which can result from this change, but TI EVMs have memories that
are large enough to require this.

NOTE: The alternate approach of enabling HIGHMEM based on SoC Kconfig
was discouraged[2] for Keystone (as example) and upstream is tending
towards a MULTIPLATFORM_TYPICAL configuration.

[1] Refs:
https://www.kernel.org/doc/Documentation/vm/highmem.txt
http://linux-mm.org/HighMemory
https://lwn.net/Articles/440221/
http://comments.gmane.org/gmane.linux.ports.arm.kernel/337727

[2] https://patchwork.kernel.org/patch/9148191/

Signed-off-by: Nishanth Menon <nm@ti.com>