ti_config_fragments/defconfig_builder: Dont generate an empty .config on interruption
Temporary empty storage file for pre-existing .config is created using
mktemp, however, when the defconfig build process is interrupted
for any reason, -f check will return true and will copy the empty
temporary file as the current .config. Such an empty .config is
invalid.

Here we have the following options:
a) Generate just the file name using --dry-run as an option to mktemp
(instead of actually creating a temporary file) - however it may have
the risk of failing for other reasons.
b) Check for valid .config prior to doing mktemp, however, we will end
up checking for empty variables
c) check if the temporary storage file is non-zero. if non-zero, then,
restore it using the '-s' check

We use (c) as an approach since the existing zero sized .config even if
exists is invalid in the first place.

Fixes: 1bfd153bc19a ("ti_config_fragments/defconfig_builder.sh: Keep developer's .config alive")
Signed-off-by: Nishanth Menon <nm@ti.com>