ti_config_fragments/connectivity.cfg: config usb net drivers to be modules
USB Net devices config options can be defined as "m"(module), as they
are not required to be available early during boot. And USB Net
options are defined as built in only for system_test purposes and this
is done in system_test.cfg already.

Hence, mark USB Net as modules to reduce boot time and Kernel image
size a bit.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>