soc: ti: opp-domain: Fix wrong scaling of VDD regulator when VBB regulator is required
Fix the copy paste typo causing VBB regulator to be never set, instead
VDD regulator is attempted to be set. This causes multiple issues
depending on Silicon type and direction of transition.

When scaling in the direction that requires ABB to be set first, indicative
voltage is first set on VDD, followed by setup of optimized voltage for
VDD. In this case, the Bias voltage remains the same as set up by boot
loader, and the VDD voltage looks right.

When scaling in the direction that requires ABB to be set last, indicative
VDD optimized voltage is setup first, followed by override with
indicative voltage on VDD. This leaves Bias voltage at the same voltage
set at boot, and VDD voltage looks broken. If we transitioned to the OPP
that bootloader setup, VBB looks right and VDD voltage looks broken.

Fix the typo. I guess we never scaled ABB.. and could only be caught
with right samples.

Fixes: 9fd88710a161 ("soc: ti: Introduce TI OPP domain driver")
Reviewed-by: Suman Anna <s-anna@ti.com>
Reported-by: Carlos Hernandez <ceh@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>