net: prueth: Fix stats before interface is started or when restarted
Initialize the DRAMs to 0 at probe time, else we see garbage
statistics till the interface is started for the first time.

As we're clearing DRAMs to 0 at emac_ndo_open() we need to save
and restore the stats whenever the interface is brought down and up
else we'll get 0 stats on every restart.

Fixes: 67c183326fcc ("net: prueth: Add TI PRUSS Ethernet driver")
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>