HACK: remoteproc/pru: allow device tree to configure GP_MUX_SEL
Add support for the property "ti,pruss-gp-mux-mode" in the device
tree that will configure a PRU core's GP_MUX_SEL value in the
PRUSS_GPCFG register.

This property is currently limited only for AM571x-IDK board and
for the PRU Ethernet usecase, due to the inherent pinmux choices
made for 6-port (4 PRU + 2 CPSW) Ethernet functionality.

Signed-off-by: Roger Quadros <rogerq@ti.com>
[s-anna@ti.com: restrict the usage for AM571x IDK PRU Ethernet usecase]
Signed-off-by: Suman Anna <s-anna@ti.com>