ti_config_fragments: Add rt based mappings to the map
The defconfig_builder script checks to ensure that the
config files for the type exist prior to displaying the
defconfig option to the user.  So on a branch where
the real_time.cfg does not exist the user will not be
presented with that option.  If the user switches to a
branch containing the real_time.cfg then the rt defconfig
options will be presented.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Acked-by: Denys Dmytriyenko <denys@ti.com>
Acked-by: Nishanth Menon <nm@ti.com>