ti_config_fragments: Provide interactive menu defconfig classification from map file
There is a need to be able to create defconfigs for
release, debug and other classes.

In order to facilitate the creation of defconfigs a dynamic
interactive menu for defconfig classification is presented
to the user.

Should a debug version be selected then a disclaimer is displayed
indicating potential performance impacts.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Acked-by: Denys Dmytriyenko <denys@ti.com>
Acked-by: Nishanth Menon <nm@ti.com>