ti_config_fragments: Remove harsh if build type is out of range
Indicating "This is not a choice try again!" is a little
harsh so make the output more meaningful.  Indicating
to the user that they need to pick between x and y
is a better approach.  This also prompts a change
to the defconfig number variable to be more explicit
about what it does.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Reviewed-by: Nishanth Menon <nm@ti.com>