ARM: AM43XX: hwmod: Fix RSTST register offset for pruss
pruss deasserting the reset was wrongly clearing bits of PWRSTCTRL
register for PER domain. In omap4_prminst_deassert_hardreset
function we alter the RSTST register value. In case of am43xx this
offset was actually pointing to PWRSTCTRL. Hence fix the RSTST register
offset value.

This was leading to setting of wrong power state values for PER
domain and causing standby failure.

Signed-off-by: Keerthy <j-keerthy@ti.com>
Tested-by: Suman Anna <s-anna@ti.com>
Acked-by: Suman Anna <s-anna@ti.com>
Acked-by: Dave Gerlach <d-gerlach@ti.com>