media: ti-vpe: vpe: Make sure vb2 buffers are not released more than once
To handle de-interlacing VPE needs to keep the previous 2 buffers
around. This is accomplished by maintaining an internal array of buffers.
When de-interlacing is active the buffer pointers will transition
between the three position in the array.

If the conversion completes or is interrupted anytime during the
processing of the first two frames, then care must be taken at
cleanup time to make sure we are not trying to free/return the
same buffer twice.

Otherwise a WARN message will be generated by the vb2 framework
for every buffers being "released/returned" more than once.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>