drm/tilcdc: Get rid of legacy dpms mechanism
commit 47bfd6c01efe639d4c44b1e3fce3816d36b23d46 upstream.

Get rid of legacy dpms mechanism. This simplifies the code quite a
bit. The old start() and stop() functions become tilcdc_crtc_enable()
and *_disable(). The functions are added with all the necessary
mechanisms from the old dpms function and they are used directly as
the crtc helper enable() and disable() callbacks.

Signed-off-by: Jyri Sarha <jsarha@ti.com>