drm/omap: omap_wb_m2m: Set v4l2 min_buffers_needed
Upon receing the STREAM_ON ioctl the V4L2 framework will check
that the current numbers of queued buffers is at least equal or greater
than 'min_buffers_needed".
For m2m the minimum # of buffers is at least 1, currently this value
is not initilialized (i.e. the current minimum is zero).

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>