drm/omap: omap_wb_cap: Set default format based on selected input
v4l2-compliance test fails when validating the format API when
G_FMT() != S_FMT(G_FMT()).

Having a single default frame format and size is not sufficient.
Default frame size needs to be valid against a specific input.
Updating the current default format to match selected input
is required to abide by the above condition.

This is not a critical error but given that v4l2 test app expect
the above behavior we need to fix it.

Signed-off-by: Benoit Parrot <bparrot@ti.com>