drm/omap: omap_wb_m2m: Make sure all VB2 buffers are sequenced properly.
v4l2-compliance test fails when testing the steaming API when
it check that the buffer sequence are not increasing.

The capture buffers were properly getting a increasing
sequence number, but the output buffers were not.

This is a benign issue however it is possible that some application
might rely on this parameters.

Signed-off-by: Benoit Parrot <bparrot@ti.com>