drm/omap: omap_wb_m2m: Remove unused ctrl handler
v4l2-compliance test fails when testing the control API.

In this driver we do not use any v4l2 control function but some
of the definition was still left around not enough to cause any
run-time issue but enough for the compliance test to show that
not all needed parts were properly provided.

Since we do not need any of this API, we just need to clean up any
reference to it. We can safely remove any partial initialization
being done.

Signed-off-by: Benoit Parrot <bparrot@ti.com>