drm/tilcdc: The dpms state should only be restored if the context is valid
The dpms state should only be restored if the context is valid. The
dpms state is only saved if LCDC was on when tilcdc_pm_suspend() was
called. This is indicated by tilcdc_drm_private ctx_valid flag. Without
this patch the lcdc driver may turn the lcdc crtc on in resume even
if it was not on before. This is a serious error since the HW is not
otherwise configured.

Fixes: 37ff3fb0dda8 ("drm/tilcdc: Restore old dpms state in pm_resume()")

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ti.com>