ARM: dts: am335x-icev2: Tweak SPI NOR partitions
The size of the SPI NOR on AM335x ICE EVM is only 8 megabytes. So the
purpose of the SPI NOR isn't necessarily to be able to completely boot
the system which would require a much bigger NOR.

So adjust the partitions so that the NOR can be used to boot up the
system to U-boot. This will allow the kernel, dtb and filesystem to
be loaded from some other storage media. Also include a misc partition
for customer specific use cases.

Signed-off-by: Franklin S Cooper Jr <fcooper@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>