ARM: dts: dra72-evm-common: Add watchdog timers for IPU1 and DSP1
The watchdog timer information has been added to the IPU1 and DSP1
remote processor device nodes for all the DRA72 EVM boards. The
data is added to the dra72-evm-common dts file since it is common
to all DRA72 EVM board variations. The following timers (same as the
timers on DRA7 EVM board) are used as the watchdog timers,
        DSP1 : GPT10
        IPU1 : GPT7 & GPT8 (one for each Cortex-M4 core)

The MPU-side drivers will use this data to initialize the watchdog
timers, and listen for any watchdog triggers. The BIOS-side code on
IPU1 and DSP1 needs to configure/refresh this timer properly to not
throw a watchdog error.

These timers can be changed or removed as per the system integration
needs, alongside appropriate equivalent changes on the firmware side.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>