media: ti-vpe: vip: Make sure frame size dont exceed scaler capacity
When scaler is to be used we need to make sure that the input frame
size does not exceed the maximum frame sizes that the scaler h/w
can handle otherwise streaming stall as the scaler cannot proceed.

The scaler buffer is limited to 2047 pixels (i.e. 11 bits) when
attempting anything larger (2048 for example) the scaler stalls
making capture stalls as well.

When checking for valid frame size in try_fmt we need to prevent
scaling when to the input size is too large for the scaler to handle.
In this case if the sensor native resolution is higher than the scaler
can handle then the frame size is set to the closest native resolution
and scaler is disabled. This allows capture to proceed.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>