ti_config_fragments: Update README to add requirements
Add a section in the README file to indicate the
basic system requirements the script needs to run.

This includes basic Linux components like bash, awk, cat
and grep versions.

Coreutils needs to be 8.1 or higher due to mktemp template
changes.  Otherwise the user will end up with the issue
described here

https://lists.gnu.org/archive/html/help-gnu-utils/2010-03/msg00001.html

Signed-off-by: Dan Murphy <dmurphy@ti.com>