ARM: dts: am4372: Set rtc as system-power-controller at board level
AM4372 had set rtc as system-power-controller in the base SoC dtsi,
however this isn't true on every board using the SoC. The IDK board
does not have the PMIC_PWR_EN line from the rtc connected to anything
so it is not possible for the rtc to power off the supplies.

Because of this, let's remove the system-power-controller flag from the
base SoC dtsi for AM4372 and instead add it to the rtc override nodes for
the boards that are able to use it to power down the system,
am437x-gp-evm and am437x-sk-evm.

Reported-by: Hongmei Gou <h-gou@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>