ARM: dts: am571x-idk-lcd: Disable PRUSS1 ethernet nodes
The AM571x IDK board can support either a 6-port Ethernet (2 CPSW
+ 4 PRUSS) with no LCD (default) or an LCD but with only 4-port
Ethernet (2 CPSW + 2 PRUSS2 Ethernet) configuration. The decision
is made based on the jumper J51 on the board. U-Boot configures the
pinmux appropriately and also chooses the correct DTB based on the
jumper setting.

PRUSS1 Ethernet ports should not be enabled on the AM571x LCD DTS
files as such, so disable the appropriate nodes inherited from the
base am571x-idk.dts file and that are enabled by default. The ethernet
aliases are also updated appropriately, and the ti,pruss-gpmux-sel
property is also deleted from corresponding PRU nodes as they are
not required in the LCD + 4-port configuration.

Tested-by: Aparna Balasubramanian <aparnab@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>