net: prueth: Rearrange data offsets to match new firmware
The new ICSS-PRU Ethernet firmware uses a new arrangement for the data
exposed to the driver in RAM due to the addition of version reporting.
The new firmware also adds support for reporting hardware queue underflow
count in addition to the existing overflow count. Add reporting of this
new information to the drivers ethtool status display.

While we are here, also update the copyright year.

Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>