drm/omap: omap_wb_m2m: Replace fourcc_to_str with a macro
commit db89da68d841 branch ti-linux-4.4.y

As a function fourcc_to_str() could potentially
produce the wrong result since it relied on a single
static char[] to present the result.

Since this is only used for debugging purposes we are
instead replacing it with explicit single byte format (aka %c)
and added a simple macro to extract a specific byte from a word.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>