drm/omap: Add top level WB support driver
commit fa82ec354c0f branch ti-linux-4.4.y

In preparation to add WB Capture mode we need to separate the
top level functions used to register with omapdrm and install
irq hooks.

As we are splitting the omapdrm registration helpers
to a new files we also need to update omap_drv to
use the more generic WB hooks as well as the
related Kconfig and Makefile changes.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	drivers/gpu/drm/omapdrm/Kconfig
	drivers/gpu/drm/omapdrm/Makefile
	drivers/gpu/drm/omapdrm/omap_drv.c