drm/omap: omap_wb_m2m: Fix buffer type g/s_selection
commit e77a97c40c1d branch ti-linux-4.4.y

v4l2-compliance test fails when validating the crop/selection API
because the buffer types expected does not match what the driver
currently advertise.

The selection ioctl had not been tested thus far and the compliance
tool quickly found an issue which was simple to fix.

Currently g_selection and s_selection are currently checking for
single plane buffer type. Although the driver only accept multi-plane
type buffer.

This error is not critical as none of our test currently attempts to
use the selection API with this driver nonetheless it is an easy fix.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>