drm/omap: omap_wb_m2m: Fix buffer clean-up issue.
commit b0c296433125 branch ti-linux-4.4.y

While testing the drievr against v4l2-compliance tool a buffer
clean-up issue was found where one buffer was never released
causing a WARN message from the vb2 frame work layer.

In order to fix this issue I re-organized the way the buffers
are handled. Buffers are only removed for the ready queue
once they have been processed instead of being saved into
the ctx structure. This makes the clean up in stop_streaming
a lot simpler especially when we are trying to abort and on-going
transaction.

As part of this code re-org the bufs_completed counters (which
was not really in-use) was removed.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>

Conflicts:
	drivers/gpu/drm/omapdrm/omap_wb.h
	drivers/gpu/drm/omapdrm/omap_wb_m2m.c