remoteproc/pruss: introduce has_no_syscfg private data field
The PRUSS_SYSCFG register is applicable only on SoCs using the
OCP interconnect methodology (so not applicable on Keystone K2G
SoCs). However, this register is also not applicable to PRUSS0
instance on AM437x SoCs even though the interconnect follows OCP.
So, make this a private data field that can be set in the instance
specific private data. The functionality around the skip_syscfg
state variable in the pruss structure is also replaced with this
field.

Signed-off-by: Suman Anna <s-anna@ti.com>