remoteproc/pruss: add a syscfg IP wrapper driver for AM437x SoCs
The AM437x SoC has two PRU-ICSS instances - PRUSS1 and PRUSS0.
The PRUSS1 instance is very similar to that of the lone PRUSS
instance on AM335x SoC. The PRUSS0 is a cut-down version of the
IP, it has less DRAM per PRU, and no Shared DRAM, a few more I/O
lines among other minor differences.

This PRUSS0 shares a single L3 OCP interconnect interface with
PRUSS1, and also shares its reset line and clocks. Any PRUSS0
functionality requires the same reset line to be deasserted and
requires the PRUSS1's PRUSS_SYSCFG register to be programmed
properly if requiring any L3 access.

This patch introduces a PRUSS syscfg IP wrapper driver that
manages the common resets, clocks and the L3 external access
specifically for AM437x SoCs. This driver is also responsible
for creating the child PRUSS devices. The code looks similar
to some of the existing code in the PRUSS driver, but this
design allows the PRUSS1 and PRUSS0 instances to be managed
independently like on AM57xx or Keystone K2G SoCs.

This is a preparatory patch, and the next patch adapts the
existing AM437x PRUSS1 node and update the PRUSS platform driver
to use this architecture.

Signed-off-by: Suman Anna <s-anna@ti.com>