remoteproc/pruss: adapt AM437x PRUSS1 node to use PRUSS wrapper driver
The AM437x PRUSS1 node is currently added directly as a child node
under the OCP node. This node is now moved under a newly created
PRUSS wrapper node with the ti,hwmods property moved to the parent
wrapper node. Only the CFG module space is mapped into the PRUSS
wrapper node to save kernel vmalloc space.

The pdata-quirks that adds the reset pdata ops to the PRUSS1 device
are also moved to the PRUSS wrapper device, and the PRUSS platform
driver is updated with new state variables that skips the management
of resets and PRUSS_CFG.SYSCFG programming, as these are handled by
the AM437x PRUSS syscfg wrapper driver.

Signed-off-by: Suman Anna <s-anna@ti.com>