TEMP: Documentation: DT: Add Keystone DSP Memory mapping device binding
Add the device tree binding document for the userspace memory mapping
driver on Keystone SoCs. This driver is used to expose either a portion
of the DDR or a portion of the on-chip RAM controlled by the Multicore
Shared Memory Controller (MSMC) to userspace for the Multi Proc Manager
(MPM) stack.

The binding status is unstable, as this might be redone completely
to accomplish this in a saner fashion.

Signed-off-by: Suman Anna <s-anna@ti.com>