mmc: host: omap_hsmmc: set 'valid' bit for the last descriptor
commit f89bdb32ed44e ("mmc: host: omap_hsmmc: Enable ADMA2")
while preparing the adma descriptor table for every mmc data
request did not set the 'valid' bit for the last descriptor or
end of descriptor.
"Table 25-14. Additional Parameters of a Descriptor Line" in
http://www.ti.com/lit/ug/spruhz6g/spruhz6g.pdf specifies that
'valid' bit should be set to indicate that the particular
descriptor line is effective (irrespective of whether the
descriptor contains valid address).
Fix it here by setting 'valid' bit in the last descriptor.

Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>