drm/omap: omap_wb_m2m: In error cases we need to terminate current transaction
As part of a previous code cleanup the code to terminates the current
transaction was removed from the job_abort() helper function.
Since in omap_wb_m2m we do not have a free running IRQ function we cannot
schedule a clean to happen when the IRQ fires because in case where a job
does not start no IRQ will be seen.

So to work around this the current transaction needs to be cleaned-up or
terminated while inside the job_abort() scope.
Otherwise you end-up with a zombie process waiting for an I/O that will
never come.

Signed-off-by: Benoit Parrot <bparrot@ti.com>