ti_config_fragments: defconfig_map: Update defconfig_map debug entries
The ti_sdk_dbg types defined in the defconfig map were
not clear and seemed to hide the fact that the classification
was a debug defconfig build.

Change the convention to remove the _dbg_ and _release from the
debug classifications and add _debug to the type.  This will clearly
indicate that the type is a debug configuration.

Signed-off-by: Dan Murphy <dmurphy@ti.com>
Acked-by: Nishanth Menon <nm@ti.com>