ARM: dts: am57xx-evm-reva3: fix typo in mmc2 pinmux offset
A terrible typo resulted in pinctrl register offsets for
mmc2 being written as 0x37xx instead of 0x34xx as they should
be. So offset 0x3498 was written as 0x3798.

And this was copied over to entries for different mmc2 modes.

This was not caught so far because U-Boot or ROM settings
still work. However, once MMC3 support for WLAN was added, it
resulted in pin clash warnings.

The offsets are correct in am57xx-beagle-x15.dts file and so
only Rev A3 of AM57xx GP EVM is affected.

Tested by read/write tests on eMMC partition on AM57x GP EVM
Rev A3.

Fixes: 7599de12482b ("ARM: dts: am57xx-evm-reva3: enable higher speed MMC1/2 modes")
Reported-by: Yan Liu <yan-liu@ti.com>
Reported-by: Eyal Reizer <eyalr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>