ARM: dts: K2G: add keystone-k2g-evm-lcd.dts
K2G EVM has HDMI output, but an LCD can be installed as a separate
add-on. The officially supported panel is Newhaven's
NHD-4.3-480272EF-ATXL.

This patch adds a new .dts file keystone-k2g-evm-lcd.dts, which includes
the base k2g evm dts file. As LCD and HDMI are exclusive, the HDMI
device is disabled in this new .dts file.

The LCD has edt-ft5306 touchscreen on i2c1 bus, which is also added.

K2G is still missing full PWM support, so backlight is not added yet.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>