net: netcp: ethss: add 10gbe serdes lane receive recovery
When the remote device that the 10gbe device is connected
to is restarted for whatever reason, the receive direction
of a 10gbe serdes lane will need to be reset in order to
sync up again with the remote device's transmit pattern.

This patch adds support of such a 10gbe serdes lane receive
recovery.  When 10gbe driver detects a serdes lane down, it
will start a periodic attempt to recover the receive direction
of that lane.

This patch also adopts the remodelling of one SerDes lane
as a separate PHY device. Hene phy_init() is invoked per
each ethernet interface.

The changes are needed because of the recent update to the
SerDes configuration.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>