media: ti-vpe: vip: Add csc support
Adding support for the colorspace converter unit.
Since there is only one csc unit shared between 2 ports (A & B),
we implement a simple resource allocation mechanism to only attempt
to allocate the csc unit if the requested colorspace is not
already natively supported by the sub device.
The csc unit will only be allocated if needed during a S_FMT
ioctl. And freed either when the device node is closed or
on subsequent S_FMT ioctl call if no longer needed.

Signed-off-by: Benoit Parrot <bparrot@ti.com>