drm/omap: displays: encoder-sii9022: Clean up hpd and interrupt handling
When the driver is polling there is no need to check the interrupt events
in the status register and there is no need to clear them either. We do
not have interrupts enabled.
If we are using the interrupt for HPD/RXsense detection then we need to
ack the interrupts in the status register.

Separate the two code path to handle polling and interrupt based detection.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ti.com>