media: ti-vpe: vip: Disable IRQs on VIP FIFO overflow events
Since enabling PARSER IRQ, when any of the overflow events
are triggered it causes a system lock up because of the event
will just keep coming until the faulty condition is cleared.

Until we implement overflow event handling it is better to just
disable the failing port.

This condition would sometime happen before we enabled the Parser
IRQs but then it would only manifest itself as as the port
appearing to be locked-up.

Signed-off-by: Benoit Parrot <bparrot@ti.com>