TEMP: soc: ti: keystone_dsp_mem: create sysfs entries for memories
The keystone-dsp-mem driver is enhanced to provide sysfs entries to
allow userspace to read the address and size of supported DDR and
Multicore Shared Memory (MSM) RAM memories that are exposed to
userspace. This sysfs logic provides an agnostic way of presenting
the supported memories irrespective of how the driver acquires the
memories.

Each supported memory region is represented by its own directory,
and are created under the dspmem misc device. The directories can
be accessed under the /sys/class/misc/dspmem/ path.

Signed-off-by: Suman Anna <s-anna@ti.com>