media: ti-vpe: vip: Eliminate an IRQ race condition on start_streaming
Since we are using the LIST_COMPLETE interrupt has our only signal
that a frame capture is complete.
However this is not true for the first two buffer descriptors as
the interrupt merely means that the descriptors have been accepted
by VPDMA. The third buffer descriptor will stall in the VPDMA pipe
until the first buffer is actually received and written to memory.

So far the way we have been handling this was to always skip the
first two LIST_COMPLETE IRQs in the ISR.
Those first two LIST_COMPLETE events exposes a race condition
which is affected by careful call sequence and/or debug level.

Stress testing has shown that on occasion by the time the IRQs were
enabled the first two LIST_COMPLETE events might have been missed.
This in turn causing a stall capture.

In order to work around this I decided to pre-load the VPDMA fifo
in a sequential manner before enabling the IRQs.

So on startup we are therefore assured that given a capture rate
of 60(30) fps we have at least 16.6(33.3) ms before the first IRQ
comes in, instead of a few microseconds.

Signed-off-by: Benoit Parrot <bparrot@ti.com>