ARM: dts: k2g: add functional clock to serdes DT node
commit 6d85dc375571 ("phy: keystone: updates to 10gbe gbe pcie
serdes configurations") has moved the SerDes h/w initialization
to probe from phy_init(). The SerDes is an integral part of the
peripheral device and depends on the same functional clock as
the peripheral. This requires that the peripheral clock is
added to the DT node of serdes device as well. This patch adds
functional clock to pcie serdes DT node for K2G. Prior to this
change, the peripheral clock was enabled by the corresponding
peripheral device driver before call to phy_init() and thus
no functional clock was added to serdes DT node.

While at it also rename the pcie serdes dt node naming to be in
sync with other peripheral serdes nodes such as gbe and xgbe.

Fixes: 762fc5d069b8 ("ARM: dts: keystone: update SerDes bindings
for one PHY per lan")
Tested-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>