clk: TI_SCI: Fix determine_rate API usage
The determine_rate function should set the new rate in the rate field
of rate request and return 0 on success, negative value on error. The
function was instead not modifying rate and returning the frequency
on success and 0 on error. This sort of worked since most frequencies
turn out to not be negative values, but failed when the user later
tried to set the frequency to the exact rate value, which may not be
possible. See rate field of clk_rate_request in clk-provider.h:

 * @rate:               Requested clock rate. This field will be adjusted by
 *                      clock drivers according to hardware capabilities.

This error manifested as a DSS clock rate request for 49.9MHz, for which
the PMMC returned that it could set 49.875MHz. The function returned a
non-negative value (49875000) but did not modify rate. The clock API
then requested a frequency of exactly 49.9MHz, which then fails.

Signed-off-by: Russ Dill <Russ.Dill@gmail.com>