ARM: dts: keystone-k2g-ice: Add a common DSP shared memory CMA pool
A common CMA memory pool reserved memory node is added, and is
attached to the DSP node through the 'memory-region' property on
the K2G ICE board. This area will be used for allocating virtio
rings and buffers. The common node allows the DSP Memory Protection
and Address Extension (MPAX) module to be configured properly for
the DSP processor.

Signed-off-by: Suman Anna <s-anna@ti.com>