drm/omap: displays: Remove panel-tlc59108
Remove panel-tlc59108 that was a quick hack to get the tlc59108 leds
controller powered and enabled LG lp101wx2 panel and backlight
working. After adding the gpio support to leds-tlc591xx driver it is
possible to use leds-tlc591xx, led-backlight, and panel-dpi to get the
LG panel working in slightly more elegant way.

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ti.com>