media: ti-vpe: vpe: HACK handle single fd multi-plane DMABUF
Special case when dealing with DMABUF buffers coming from the
g-streamer vpe-plugin. These DMABUF buffers are setup to look
like they have 2 planes but in reality both plane fd are the same.
So they are actually single plane buffer.

In order to compute the buffer offset properly we need to treat
this as a special case. Since this is a hack at best we also
include a WARN_ONCE() message so the user space side can be fixed.

Signed-off-by: Benoit Parrot <bparrot@ti.com>