ti_config_fragments: increase CMA size for AM3/4 & K2G
AM3 BeagleBone Black, AM4 EVM and K2G EVM all have HDMI output, and
require contiguous memory from CMA for the HDMI buffers. The current CMA
size is not enough for many common use cases.

This patch increases the CMA memory to 48M in *_only.cfg fragments. This
is far from optimal, as even boards without any displays will get
increased CMA area.

Preferably I would want to increase the default CMA memory area from a
board file. This can be done, but results in "cma=" kernel parameter to
be ignored, removing the user's option to change the CMA, which is not
acceptable.

Alternatively, a separate CMA area for display driver could be created,
but this requires more work and testing, and is also not optimal as then
that CMA area is out of other drivers' use.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>